decouple baseline code in the graph submodule
