+++ .gitignore
@@ -7,3 +7,4 @@ cache
 .DS_Store
 **/*.egg-info
 notebooks
+venv
