+++ graph/models/baseline.py
@@ -1,0 +1,44 @@
+import torch
+from torch import nn as nn
+from torch.nn import functional as F
+
+
+class GNN(nn.Module):
+    """
+    Simple message passing model that consists of 2 message passing layers
+    and the sum aggregation function
+    """
+    def __init__(self, input_dim, hidden_dim, dropout, n_class):
+        super(GNN, self).__init__()
+        self.fc1 = nn.Linear(input_dim, hidden_dim)
+        self.fc2 = nn.Linear(hidden_dim, hidden_dim)
+        self.fc3 = nn.Linear(hidden_dim, hidden_dim)
+        self.fc4 = nn.Linear(hidden_dim, n_class)
+        self.bn = nn.BatchNorm1d(hidden_dim)
+        self.relu = nn.ReLU()
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, x_in, adj, idx):
+        # first message passing layer
+        x = self.fc1(x_in)
+        x = self.relu(torch.mm(adj, x))
+        x = self.dropout(x)
+
+        # second message passing layer
+        x = self.fc2(x)
+        x = self.relu(torch.mm(adj, x))
+
+        # sum aggregator
+        idx = idx.unsqueeze(1).repeat(1, x.size(1))
+        out = torch.zeros(torch.max(idx)+1, x.size(1)).to(x_in.device)
+        out = out.scatter_add_(0, idx, x)
+
+        # batch normalization layer
+        out = self.bn(out)
+
+        # mlp to produce output
+        out = self.relu(self.fc3(out))
+        out = self.dropout(out)
+        out = self.fc4(out)
+
+        return F.log_softmax(out, dim=1)
