+++ structure_baseline.py
@@ -1,159 +1,24 @@
 import csv
-import pickle
 import time
 import numpy as np
 import scipy.sparse as sp
 
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 from torch import optim
-from config import CACHE_DIR, DATA_DIR, OUTPUT_DIR
+from config import OUTPUT_DIR
+from graph.data import load_graph_data, normalize_adjacency, sparse_mx_to_torch_sparse_tensor, train_test_split
+from graph.models.baseline import GNN
 
-def load_data(): 
-    """
-    Function that loads graphs
-    """
-    try:
-        with open(f'{CACHE_DIR}/adj.pickle', 'rb') as f:
-            adj = pickle.load(f)
-        with open(f'{CACHE_DIR}/features.pickle', 'rb') as f:
-            features = pickle.load(f)
-        with open(f'{CACHE_DIR}/edge_features.pickle', 'rb') as f:
-            edge_features = pickle.load(f)
-        print(f"Loading cached features from {CACHE_DIR}")
-
-    except FileNotFoundError:
-        print("Calculating features and adjacency matrix")
-        graph_indicator = np.loadtxt(f"{DATA_DIR}/graph_indicator.txt", dtype=np.int64)
-        _, graph_size = np.unique(graph_indicator, return_counts=True)
-
-        edges = np.loadtxt(f"{DATA_DIR}/edgelist.txt", dtype=np.int64, delimiter=",")
-        edges_inv = np.vstack((edges[:,1], edges[:,0]))
-        edges = np.vstack((edges, edges_inv.T))
-        s = edges[:,0]*graph_indicator.size + edges[:,1]
-        idx_sort = np.argsort(s)
-        edges = edges[idx_sort,:]
-        edges,idx_unique =  np.unique(edges, axis=0, return_index=True)
-        A = sp.csr_matrix((np.ones(edges.shape[0]), (edges[:,0], edges[:,1])), shape=(graph_indicator.size, graph_indicator.size))
-
-        x = np.loadtxt(f"{DATA_DIR}/node_attributes.txt", delimiter=",")
-        edge_attr = np.loadtxt(f"{DATA_DIR}/edge_attributes.txt", delimiter=",")
-        edge_attr = np.vstack((edge_attr,edge_attr))
-        edge_attr = edge_attr[idx_sort,:]
-        edge_attr = edge_attr[idx_unique,:]
-
-        adj = []
-        features = []
-        edge_features = []
-        idx_n = 0
-        idx_m = 0
-        for i in range(graph_size.size):
-            adj.append(A[idx_n:idx_n+graph_size[i],idx_n:idx_n+graph_size[i]])
-            edge_features.append(edge_attr[idx_m:idx_m+adj[i].nnz,:])
-            features.append(x[idx_n:idx_n+graph_size[i],:])
-            idx_n += graph_size[i]
-            idx_m += adj[i].nnz
-
-        with open(f'{CACHE_DIR}/adj.pickle', 'wb') as f:
-            adj = pickle.dump(adj, f)
-        with open(f'{CACHE_DIR}/features.pickle', 'wb') as f:
-            features = pickle.dump(features, f)
-        with open(f'{CACHE_DIR}/edge_features.pickle', 'wb') as f:
-            edge_features = pickle.dump(edge_features, f)
-        print("Features and adjacency matrix cached for future use.")
-
-    return adj, features, edge_features
-
-def normalize_adjacency(A):
-    """
-    Function that normalizes an adjacency matrix
-    """
-    n = A.shape[0]
-    A += sp.identity(n)
-    degs = A.dot(np.ones(n))
-    inv_degs = np.power(degs, -1)
-    D = sp.diags(inv_degs)
-    A_normalized = D.dot(A)
-
-    return A_normalized
-
-def sparse_mx_to_torch_sparse_tensor(sparse_mx):
-    """
-    Function that converts a Scipy sparse matrix to a sparse Torch tensor
-    """
-    sparse_mx = sparse_mx.tocoo().astype(np.float32)
-    indices = torch.from_numpy(np.vstack((sparse_mx.row, sparse_mx.col)).astype(np.int64))
-    values = torch.from_numpy(sparse_mx.data)
-    shape = torch.Size(sparse_mx.shape)
-    return torch.sparse.FloatTensor(indices, values, shape)
-
-
-class GNN(nn.Module):
-    """
-    Simple message passing model that consists of 2 message passing layers
-    and the sum aggregation function
-    """
-    def __init__(self, input_dim, hidden_dim, dropout, n_class):
-        super(GNN, self).__init__()
-        self.fc1 = nn.Linear(input_dim, hidden_dim)
-        self.fc2 = nn.Linear(hidden_dim, hidden_dim)
-        self.fc3 = nn.Linear(hidden_dim, hidden_dim)
-        self.fc4 = nn.Linear(hidden_dim, n_class)
-        self.bn = nn.BatchNorm1d(hidden_dim)
-        self.relu = nn.ReLU()
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, x_in, adj, idx):
-        # first message passing layer
-        x = self.fc1(x_in)
-        x = self.relu(torch.mm(adj, x))
-        x = self.dropout(x)
-
-        # second message passing layer
-        x = self.fc2(x)
-        x = self.relu(torch.mm(adj, x))
-        
-        # sum aggregator
-        idx = idx.unsqueeze(1).repeat(1, x.size(1))
-        out = torch.zeros(torch.max(idx)+1, x.size(1)).to(x_in.device)
-        out = out.scatter_add_(0, idx, x)
-        
-        # batch normalization layer
-        out = self.bn(out)
-
-        # mlp to produce output
-        out = self.relu(self.fc3(out))
-        out = self.dropout(out)
-        out = self.fc4(out)
-
-        return F.log_softmax(out, dim=1)
 
 # Load graphs
-adj, features, edge_features = load_data()
+adj, features, edge_features = load_graph_data()
 
 # Normalize adjacency matrices
 adj = [normalize_adjacency(A) for A in adj]
 
-
 # Split data into training and test sets
-adj_train = list()
-features_train = list()
-y_train = list()
-adj_test = list()
-features_test = list()
-proteins_test = list()
-with open(f'{DATA_DIR}/graph_labels.txt', 'r') as f:
-    for i, line in enumerate(f):
-        t = line.split(',')
-        if len(t[1][:-1]) == 0:
-            proteins_test.append(t[0])
-            adj_test.append(adj[i])
-            features_test.append(features[i])
-        else:
-            adj_train.append(adj[i])
-            features_train.append(features[i])
-            y_train.append(int(t[1][:-1]))
+adj_train, features_train, y_train, adj_test, features_test, proteins_test = train_test_split(adj, features)
 
 # Initialize device
 device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
